refactor : move interfaces to domain layer
