+++ src/api/presentation/controllers/list-all-files-controller.ts
@@ -4,7 +4,7 @@ import {
   HttpResponse,
 } from '../../../core/application/domain/controller'
 import { HttpStatusCode } from '../../../core/application/domain/http-status'
-import { IListFilesUseCase } from '../../../core/application/usecases/list-files-use-case'
+import { IListFilesUseCase } from '../../../core/application/domain/usecases/list-all-files'
 
 export class ListFilesController implements Controller {
   constructor(readonly listFilesUseCase: IListFilesUseCase) {}
