+++ src/core/application/domain/usecases/delete-file.ts
@@ -1,0 +1,3 @@
+export interface IDeleteFileUseCase {
+  invoke(file_id: string): Promise<void>
+}
