+++ src/core/application/domain/usecases/list-all-files.ts
@@ -1,0 +1,5 @@
+import { _Object } from '@aws-sdk/client-s3'
+
+export interface IListFilesUseCase {
+  invoke(): Promise<_Object[]>
+}
