+++ src/core/application/domain/usecases/list-file-by-id.ts
@@ -1,0 +1,5 @@
+import { File } from '../file'
+
+export interface IListFileByIdUseCase {
+  invoke(file_id: string): Promise<File>
+}
