+++ src/core/application/domain/usecases/upload-file.ts
@@ -1,0 +1,5 @@
+import { PutObjectCommandOutput } from '@aws-sdk/client-s3'
+
+export interface IUploadFileUseCase {
+  invoke(file: Express.Multer.File): Promise<PutObjectCommandOutput>
+}
