+++ src/core/application/domain/usecases/upload-files.ts
@@ -1,0 +1,3 @@
+export interface IUploadFilesUseCase {
+  invoke(files: Express.Multer.File[]): Promise<void>
+}
